Client: refactor transaction component, use it for update and create
